Revert "Feature/refactoring"
