+++ src/entities/userInfo/index.ts
@@ -1,3 +1,2 @@
 export { userApi } from './api/userApi'
 export { type UserInfo } from './api/types'
-export { useUserInfoStore } from './model/userInfoStore'
