--- src/views/login/index.ts
@@ -1,1 +1,0 @@
-export { LoginPage } from './ui/page'
