--- src/views/onboarding/index.ts
@@ -1,1 +1,0 @@
-export { OnboardingPage } from './ui/page'
